Refactor first attempt on async call
Had to refactor big parts of the code as the `Async.Promise` module
wasn't working together with `Web.HTTP` as I initially expected.

Because of the way `Web.HTTP` implements its `#request` function it's
not possible to make it asynchronous. In this case we want to pass a
command to an external process so we don't choke Vim. However the
implementation of the `#request` method executes the call without
passing it to another thread, with no hook for us to interpret and
change the original behaviour.